refactor display, introduce monitor
Before this commit the `WindowManager` would own the `Display` it resides on, now the `Display` owns the `WindowManager`.
I realized the original structure was going to cause problems while
introducing the `NativeMonitor` related stuff.

In the first nog version the `Display` would both have cross platform
logic and logic related to os specific stuff. In the new version the
os specific logic is extracted into `Monitor`.

For example the `MonitorId` on Windows is the HMONITOR of each monitor,
while the `DisplayId` is probably going to be GUID of the phyisal
monitor (or at least something which is persistent through
unplugging monitors).