Displays are no longer hardcoded
In the beginning I used to just have a hardcoded primary display for
testing purposes, but now it gets fetched dynamically using the winapi.

On windows the DisplayId is now the DeviceID (String) of the display devices,
which should be unique between reboots AFAIK.

On windows the MonitorId is just the HMONITOR. It is important to know
that H* types (like HWND and HMONITOR) are process-unique, meaning that
between restarts of nog the handles change. This is also why we can't
just the HMONITOR as DisplayId, because it wouldn't be reliable.