add support for registering plugins
In the first version we used to have integrated support for plugin
management in the rust side, but after looking at neovim some more I
think having basic plugin management integrated and more complex stuff
on the user side is a better idea.

For example before you could just call the `nog.plug_install` function
to clone a git repo based on the `<USERNAME>/<REPO>` schema and so on.

Now you would have to manually clone the repo into `$CONFIGPATH/plugins`
and call the `nog.plugin_register` function with the name of the folder
or use a plugin manager from the userspace.