first attempt at replicating rules using events
Using the event handlers implemented in 81e34bd, I now managed to add a
system that somewhat mirrors the rule system in the first version for
simple cases.

The new implementation is completely written in lua, so a user could
implement his own way of doing rule matching or add a plugin that does
it better than the default.

In the lua snippet below I define a variable called `window_handlers`,
where each item has to have a `when` property and `action` property.

```lua
local window_handlers = {
  {
    -- This is responsible for controlling when the `action` is applied to the window.
    when = {
      title = "^.* - Mozilla Firefox$",
    },
    action = {
      workspace_id = 2
    }
  }
}
```

Right now `window_handlers` only has one handler which moves
a window that matches the lua pattern `^.* - Mozilla Firefox$` to the
workspace with the id `2`.

In the future we could have a similar lua based system for customizing
workspaces.

cc @ramirezmike